Changed survey_type to just be a string -- EnumProperty and EnumField were just too cumbersome when needing to convert between them and JSON.
Finished the taskqueue implementation of the data export. Had to write a
custom JSONEncoder to allow datetime encoding (ISO format). Currently
doing a less-than-ideal manual JSON writing routine, but it allows the
task to run with much lower memory overhead (it can write out a record
at a time to Cloud Storage instead of having to load all the records
into memory, and then process them into a dict and then into JSON).