Added an alarm created during setup to uninstall the extension after 120 days.
Added a listener for that alarm.

Renamed the alarms to differentiate between the notification timeout
alarm and the uninstall alarm.

Added a check to clearNotifications to only run the body if it's the
notification timeout alarm.

Moved the onAlarm.addListener for the notification timeout next to
clearNotification, where it's a little clearer.